implement search handlers

+++ neoClient.go
@@ -8,11 +8,11 @@ import (
 	"github.com/neo4j/neo4j-go-driver/neo4j"
 )
 
-type NeoClient struct{
+type NeoClient struct {
 	driver neo4j.Driver
 }
 
-func NewNeoClient(neoURL string) (*NeoClient, error){
+func NewNeoClient(neoURL string) (*NeoClient, error) {
 	neoDriver, err := neo4j.NewDriver(neoURL, neo4j.NoAuth(), func(config *neo4j.Config) {
 		config.Log = neo4j.ConsoleLogger(neo4j.INFO)
 	})
@@ -22,15 +22,15 @@ func NewNeoClient(neoURL string) (*NeoClient, error){
 	return &NeoClient{neoDriver}, nil
 }
 
-type SearchObject struct{
-	fromDate int
-	toDate int
-	isRelatedWith []string
+type SearchObject struct {
+	fromDate              int64
+	toDate                int64
+	isRelatedWith         []string
 	isDirectlyRelatedWith []string
-	limit int
+	limit                 int
 }
 
-func (nc *NeoClient) Search(sObj *SearchObject) ([]string, error){
+func (nc *NeoClient) Search(sObj *SearchObject) ([]string, error) {
 	var session neo4j.Session
 	var records []neo4j.Record
 	var err error
@@ -65,12 +65,12 @@ func (nc *NeoClient) Search(sObj *SearchObject) ([]string, error){
 			logger.Error("not found uuid for record")
 		}
 
-		results = append(results, fmt.Sprintf("%v",recordUUID))
+		results = append(results, fmt.Sprintf("%v", recordUUID))
 	}
 
 	return results, nil
 }
 
-func (nc *NeoClient) Close()error{
+func (nc *NeoClient) Close() error {
 	return nc.driver.Close()
 }
